refactor: Split out locales config from messages setup
