+++ src/components/header.tsx
@@ -9,7 +9,7 @@ import contentRoutes from "src/content-routes.json";
 import { IconChevronDown } from "src/icons/ic-chevron-down";
 import { useLocale } from "src/lib/use-locale";
 import { useResizeObserver } from "src/lib/use-resize-observer";
-import { locales } from "src/locales/locales";
+import { locales } from "src/locales/config";
 import { palette } from "src/themes/palette";
 import { F, FlagList, useFlag } from "src/utils/flags";
