--- src/locales/locales.json
@@ -1,4 +1,0 @@
-{
-  "locales": ["de", "fr", "it", "en"],
-  "defaultLocale": "en"
-}
