+++ src/pages/sunshine/[entity]/[id]/operational-standards.tsx
@@ -29,7 +29,7 @@ import {
   SunshineOperationalStandardsData,
 } from "src/domain/data";
 import { getLocalizedLabel } from "src/domain/translation";
-import { defaultLocale } from "src/locales/locales";
+import { defaultLocale } from "src/locales/config";
 
 type Props =
   | (Extract<SharedPageProps, { entity: "operator"; status: "found" }> & {
